move types and update storybook
